+++ VMDeploy.Orchestrator/changelog.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
+## ???
+
++ Upd: New-VmoVirtualMachine - Exports intended name of computer to resources
++ Upd: New-VmoVirtualMachine - Failure to add a disk will fail the deployment
++ Fix: New-VmoVirtualMachine - SubnetMask still exists in code when should instead be PrefixLength
++ Fix: New-VmoVirtualMachine - Removed error over guest config not found when not explicitly specifying a GuestConfig via Parameter
+
 ## 1.0.0 (2021-04-07)
- - New: Some Stuff
- - Upd: Moar Stuff
- - Fix: Much Stuff
+
++ Initial Release